Change ComponentImplementationBuilder into ComponentImplementationFactory so that we don't have to pass around easily-injectable objects.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=220464228
