+++ java/dagger/internal/codegen/BUILD
@@ -327,7 +327,7 @@ java_library(
         "BindingMethodValidatorsModule.java",
         "ComponentGenerator.java",
         "ComponentHjarProcessingStep.java",
-        "ComponentImplementationBuilder.java",
+        "ComponentImplementationFactory.java",
         "ComponentProcessingStep.java",
         "ComponentProcessor.java",
         "DiagnosticReporterFactory.java",
